LayoutTests/fast/dom/DOMException/stack-trace-expected.txt

PASS "stack" in e is true
PASS typeof e.stack is "string"
PASS e.stack.contains("innerFunction") is true
PASS e.stack.contains("outerFunction") is true
PASS Object.prototype.toString.call(e) is "[object DOMException]"
PASS e.name is "HierarchyRequestError"
PASS e.message is "HierarchyRequestError: DOM Exception 3"
PASS e.code is 3
PASS e.stack === 42 is true
PASS successfullyParsed is true

TEST COMPLETE
